third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/reporting/access-reporting/reporting-observer-expected.txt

This is a testharness.js-based test.
FAIL Opener COOP assert_equals: No report received. expected 1 but got 0
FAIL Openee COOP assert_equals: No report received. expected 1 but got 0
Harness: the test ran to completion.
